+++ CHANGELOG.md
@@ -5,6 +5,7 @@ AppEngine version, listed here to ease deployment and troubleshooting.
  * Bump runtimeVersion to `2026.02.03`.
  * Upgraded pana to `0.23.9`.
  * Note: `pub_worker` started to run more `pana` subprocess in gvisor sandbox (`git`, `webp`).
+ * Note: `pub_worker` started to use AOT-compiled `pub_worker_subprocess`.
 
 ## `20260203t095200-all`
  * Bump runtimeVersion to `2026.01.30`.